+++ modules/storage/ecr/app.py
@@ -31,6 +31,10 @@ repository_name = os.getenv(_param("REPOSITORY_NAME"), DEFAULT_REPOSITORY_NAME)
 image_tag_mutability = os.getenv(_param("IMAGE_TAG_MUTABILITY"), DEFAULT_IMAGE_MUTABILITY)
 lifecycle_max_image_count = os.getenv(_param("LIFECYCLE_MAX_IMAGE_COUNT"), DEFAULT_LIFECYCLE)
 lifecycle_max_days = os.getenv(_param("LIFECYCLE_MAX_DAYS"), DEFAULT_LIFECYCLE)
+removal_policy = os.getenv("SEEDFARMER_REMOVAL_POLICY", "DESTROY")
+
+if removal_policy not in ["DESTROY", "RETAIN"]:
+    raise ValueError("The only REMOVAL_POLICY values accepted are 'DESTROY' and 'RETAIN' ")
 
 app = cdk.App()
 stack = EcrStack(
@@ -40,6 +44,7 @@ stack = EcrStack(
     image_tag_mutability=image_tag_mutability,
     lifecycle_max_image_count=lifecycle_max_image_count,
     lifecycle_max_days=lifecycle_max_days,
+    removal_policy=removal_policy,
     env=environment,
 )
 metadata = {